Refactor types a bit
