+++ packages/craftcms-cp/src/components/action-item/action-item.ts
@@ -1,22 +1,13 @@
 import {html, LitElement, nothing} from 'lit';
 import {property, state} from 'lit/decorators.js';
 import styles from './action-item.styles.js';
-import {
-  type AsyncState,
-  AsyncStates,
-  Variant,
-  type VariantKey,
-} from '@src/types';
+import {type AsyncState, AsyncStates} from '@src/types';
 import variantsStyles from '@src/styles/variants.styles';
 import {classMap} from 'lit/directives/class-map.js';
 
 import '../shortcut/shortcut.js';
-import {
-  type ActionFeedback,
-  type BaseAction,
-  type FeedbackData,
-  runAction,
-} from '@src/actions';
+import {type ActionFeedback, type BaseAction, type FeedbackData, runAction,} from '@src/actions';
+import {Variant, type VariantKey} from '@src/constants/variants';
 
 /**
  * @summary Either a link or button typically used in a menu.
