+++ packages/craftcms-cp/src/constants/variants.ts
@@ -1,0 +1,12 @@
+export const Variant = {
+  Default: 'default',
+  Success: 'success',
+  Warning: 'warning',
+  Danger: 'danger',
+  Info: 'info',
+} as const;
+
+export const variants = Object.values(Variant);
+
+export type VariantKey = (typeof Variant)[keyof typeof Variant];
+export type VariantValue = typeof variants;
